Changes to characters screen and creation of games tab and mimic game
Character screen: Now if the user fetched and marked a show's characters, the other seasons of the same show will appear as marked.

Games screen: Added a stack navigation to select different types of games (only one for the moment).

Mimic Game: Created the game where you can play with a single device. In the setup you can add participants and in the game a random character from the user's list will be selected to mimic. Puntuation and rounds implemented.